+++ airflow/api_fastapi/core_api/routes/public/variables.py
@@ -36,15 +36,17 @@ from airflow.api_fastapi.core_api.datamodels.variables import (
     VariableBody,
     VariableBulkActionResponse,
     VariableBulkBody,
-    VariableBulkCreateAction,
-    VariableBulkDeleteAction,
     VariableBulkResponse,
-    VariableBulkUpdateAction,
     VariableCollectionResponse,
     VariableResponse,
     VariablesImportResponse,
 )
 from airflow.api_fastapi.core_api.openapi.exceptions import create_openapi_http_exception_doc
+from airflow.api_fastapi.core_api.services.public.variables import (
+    handle_bulk_create,
+    handle_bulk_delete,
+    handle_bulk_update,
+)
 from airflow.models.variable import Variable
 
 variables_router = AirflowRouter(tags=["Variable"], prefix="/variables")
@@ -241,106 +243,6 @@ def import_variables(
         import_count=len(import_keys),
         created_variable_keys=list(create_keys),
     )
-
-
-def categorize_keys(session, keys: set) -> tuple[set, set]:
-    """Categorize the given keys into matched_keys and not_found_keys based on existing keys."""
-    existing_keys = {variable for variable in session.execute(select(Variable.key)).scalars()}
-    matched_keys = existing_keys & keys
-    not_found_keys = keys - existing_keys
-    return matched_keys, not_found_keys
-
-
-def handle_bulk_create(
-    session, action: VariableBulkCreateAction, results: VariableBulkActionResponse
-) -> None:
-    """Bulk create variables."""
-    to_create_keys = {variable.key for variable in action.variables}
-    matched_keys, not_found_keys = categorize_keys(session, to_create_keys)
-
-    try:
-        if action.action_if_exists == "fail" and matched_keys:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail=f"The variables with these keys: {matched_keys} already exist.",
-            )
-        elif action.action_if_exists == "skip":
-            create_keys = not_found_keys
-        else:
-            create_keys = to_create_keys
-
-        for variable in action.variables:
-            if variable.key in create_keys:
-                Variable.set(
-                    key=variable.key, value=variable.value, description=variable.description, session=session
-                )
-                results.success.append(variable.key)
-
-    except HTTPException as e:
-        results.errors.append({"error": f"{e.detail}", "status_code": e.status_code})
-
-
-def handle_bulk_update(
-    session, action: VariableBulkUpdateAction, results: VariableBulkActionResponse
-) -> None:
-    """Bulk Update variables."""
-    to_update_keys = {variable.key for variable in action.variables}
-    matched_keys, not_found_keys = categorize_keys(session, to_update_keys)
-
-    try:
-        if action.action_if_not_exists == "fail" and not_found_keys:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"The variables with these keys: {not_found_keys} were not found.",
-            )
-        elif action.action_if_not_exists == "skip":
-            update_keys = matched_keys
-        else:
-            update_keys = to_update_keys
-
-        for variable in action.variables:
-            if variable.key in update_keys:
-                old_variable = session.scalar(select(Variable).filter_by(key=variable.key).limit(1))
-                VariableBody(**variable.model_dump())
-                data = variable.model_dump(exclude={"key"}, by_alias=True)
-
-                for key, val in data.items():
-                    setattr(old_variable, key, val)
-                results.success.append(variable.key)
-
-    except HTTPException as e:
-        results.errors.append({"error": f"{e.detail}", "status_code": e.status_code})
-
-    except ValidationError as e:
-        results.errors.append({"error": f"{e.errors()}"})
-
-
-def handle_bulk_delete(
-    session, action: VariableBulkDeleteAction, results: VariableBulkActionResponse
-) -> None:
-    """Bulk delete variables."""
-    to_delete_keys = set(action.keys)
-    matched_keys, not_found_keys = categorize_keys(session, to_delete_keys)
-
-    try:
-        if action.action_if_not_exists == "fail" and not_found_keys:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"The variables with these keys: {not_found_keys} were not found.",
-            )
-        elif action.action_if_not_exists == "skip":
-            delete_keys = matched_keys
-        else:
-            delete_keys = to_delete_keys
-
-        for key in delete_keys:
-            existing_variable = session.scalar(select(Variable).where(Variable.key == key).limit(1))
-            if existing_variable:
-                session.delete(existing_variable)
-                results.success.append(key)
-
-    except HTTPException as e:
-        results.errors.append({"error": f"{e.detail}", "status_code": e.status_code})
 
 
 @variables_router.patch("")
